rename schema_prepare to prepare_database
