+++ synapse/app/homeserver.py
@@ -36,7 +36,7 @@ if __name__ == '__main__':
 
 from synapse.storage.engines import create_engine, IncorrectDatabaseSetup
 from synapse.storage import are_all_users_on_domain
-from synapse.storage.schema_prepare import UpgradeDatabaseException
+from synapse.storage.prepare_database import UpgradeDatabaseException
 
 from synapse.server import HomeServer
